+++ src/components/UI/Home_Header.jsx
@@ -1,0 +1,10 @@
+const Header = () => {
+  return (
+    <header id="home-header">
+      <button id="signUp-btn">sign up!</button>
+      <button id="login-btn">log in!</button>
+    </header>
+  );
+};
+
+export { Header };
