+++ src/vanillaJsFiles/heros.js
@@ -1,4 +1,4 @@
-import { Hero } from "../utils.js";
+import { Hero } from "./constructor.js";
 const deathImg = "https://i.imgur.com/jEiDFYT.png";
 const bulwark = new Hero(
   "Bulwark", //name
